Give React SPA frontend specific things a home
