+++ src/argus/auth/views.py
@@ -17,9 +17,7 @@ from .serializers import (
     EmptySerializer,
     RefreshTokenSerializer,
     UserSerializer,
-    AuthMethodSerializer,
 )
-from .utils import get_authentication_backend_name_and_type
 
 
 class ObtainNewAuthToken(ObtainAuthToken):
@@ -66,21 +64,6 @@ class BasicUserDetail(generics.RetrieveAPIView):
     permission_classes = [IsAuthenticated]
     serializer_class = BasicUserSerializer
     queryset = User.objects.all()
-
-
-@extend_schema_view(
-    get=extend_schema(
-        responses=AuthMethodSerializer,
-    ),
-)
-class AuthMethodListView(APIView):
-    http_method_names = ["get", "head", "options", "trace"]
-    permission_classes = [AllowAny]
-
-    def get(self, request, *args, **kwargs):
-        data = get_authentication_backend_name_and_type(request=request)
-
-        return Response(data)
 
 
 @extend_schema_view(
